test_requirements.txt
# ==============================================================
# Packages needed for running the tests. Needed by contributors.
# ==============================================================
coverage==4.0.3
django-coverage==1.2.4
flake8==2.5.4
ipdb==0.9.0
django-nose==1.4.3
mixer==5.5.6
fabric3
beautifulsoup4
tox
